requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.2
Django==3.0
django-bootstrap4==1.1.1
django-taggit==1.2.0
mysqlclient==1.4.6
python-decouple==3.3
pytz==2019.3
soupsieve==1.9.5
sqlparse==0.3.0
